숙소 상세보기에서 리뷰 팝업

+++ lib/data/models/review.dart
@@ -1,5 +1,5 @@
 class Review {
-  final int rating; // 별점
+  final double rating; // 별점
   final String comment; // 코멘트
 
   Review({
